fix: 즐겨찾기 관련 버그 수정
